Fixed example2
- Have to use float because it that's what is descibed in the map file.
  We are skipping the lmap which would adapt firmware types to the application type, so the type has to match
  The type has changed because it was always wrongly set to double for all floating types data, although the catalogue suggests float. This fix requires the test to be adapted.

+++ example2/demoApp2.cc
@@ -6,12 +6,12 @@ namespace ctk = ChimeraTK;
 
 struct Controller : public ctk::ApplicationModule {
   using ctk::ApplicationModule::ApplicationModule;
-  ctk::ScalarPollInput<double> sp{this, "temperatureSetpoint", "degC", "Description"};
-  ctk::ScalarPushInput<double> rb{this, "temperatureReadback", "degC", "..."};
-  ctk::ScalarOutput<double> cur{this, "heatingCurrent", "mA", "..."};
+  ctk::ScalarPollInput<float> sp{this, "temperatureSetpoint", "degC", "Description"};
+  ctk::ScalarPushInput<float> rb{this, "temperatureReadback", "degC", "..."};
+  ctk::ScalarOutput<float> cur{this, "heatingCurrent", "mA", "..."};
 
   void mainLoop() {
-    const double gain = 100.0;
+    const float gain = 100.0;
     while(true) {
       readAll(); // waits until rb updated, then reads sp
